refactor(v3.2.2): 抽 692 行 _viz_xxx + DIM_VIZ_RENDERERS 到 lib/report/dim_viz
- 19 个 _viz_xxx(raw) 维度特化可视化函数
- DIM_VIZ_RENDERERS dispatch dict
- _score_class helper
- 行数: 2400 → 1708 行 (-29% · 累计从 2964 -42%)
- 332 tests 全过 · e2e assemble 零影响
